B2B UI Cleanup
