+++ Sources/StytchUI/StytchB2BUIClient/Shared/StateManagement/B2BAuthenticationManager.swift
