+++ Sources/StytchUI/StytchB2BUIClient/ViewControllers/PasswordForgotViewController/PasswordForgotViewController.swift
@@ -44,6 +44,10 @@ final class PasswordForgotViewController: BaseViewController<PasswordForgotState
             stackView.arrangedSubviews.map { $0.widthAnchor.constraint(equalTo: stackView.widthAnchor) }
         )
 
+        NSLayoutConstraint.activate([
+            continueButton.heightAnchor.constraint(equalToConstant: .buttonHeight),
+        ])
+
         setupEmailInput(input: emailInput)
     }
